feature/calculator/service without interface
